+++ src/pages/admin/index.tsx
@@ -1,28 +1,14 @@
 import React from 'react'
 import Layout from '@/components/admin/Layout'
 import { Flex, Heading, Spacer } from '@chakra-ui/layout'
-import CreateTableModal from '@/components/admin/CreateTableModal'
-import TablesList from '@/components/admin/TablesList'
+import Login from '@/pages/admin/auth/index'
 
-const Header = () => {
+const InitialPage = () => {
   return (
-    <Flex sx={{ width: '100%' }}>
-      <Heading>Mesas</Heading>
-      <Spacer />
-      <CreateTableModal />
+    <Flex direction="column" sx={{ width: 'full', padding: 4 }}>
+      <Login />
     </Flex>
   )
 }
 
-const MesasPage = () => {
-  return (
-    <Layout>
-      <Flex direction="column" sx={{ width: 'full', padding: 4 }}>
-        <Header />
-        <TablesList />
-      </Flex>
-    </Layout>
-  )
-}
-
-export default MesasPage
+export default InitialPage
