+++ src/pages/admin/mesas/index.tsx
@@ -1,0 +1,28 @@
+import React from 'react'
+import Layout from '@/components/admin/Layout'
+import { Flex, Heading, Spacer } from '@chakra-ui/layout'
+import CreateTableModal from '@/components/admin/CreateTableModal'
+import TablesList from '@/components/admin/TablesList'
+
+const Header = () => {
+  return (
+    <Flex sx={{ width: '100%' }}>
+      <Heading>Mesas</Heading>
+      <Spacer />
+      <CreateTableModal />
+    </Flex>
+  )
+}
+
+const MesasPage = () => {
+  return (
+    <Layout>
+      <Flex direction="column" sx={{ width: 'full', padding: 4 }}>
+        <Header />
+        <TablesList />
+      </Flex>
+    </Layout>
+  )
+}
+
+export default MesasPage
